frontend/pyproject.toml
[project]
name = "MLDeployLab-frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "streamlit>=1.46.1",
]
